Extract some LLVM stuff into utility module
