+++ lib/targets/llvm.js
@@ -11,8 +11,12 @@ var _            = require('lodash'),
     Errors       = require('../errors'),
     ICE          = Errors.InternalCompilerError,
     // Target information and setup
-    target       = require('./llvm/target')
+    target       = require('./llvm/target'),
+    util         = require('./llvm/util')
 
+var isLastInstructionTerminator = util.isLastInstructionTerminator,
+    compileTruthyTest           = util.compileTruthyTest,
+    assertInstanceOf            = util.assertInstanceOf
 
 var NativeFunction    = require('./llvm/native-function'),
     NativeObject      = require('./llvm/native-object'),
@@ -40,6 +44,13 @@ var TypeOf            = LLVM.Library.LLVMTypeOf,
     PointerTypeKind   = GetTypeKind(Int8PtrType)
 
 function Context () {}
+
+function BlockContext (ctx, parentFn, block, slots) {
+  this.ctx   = ctx
+  this.fn    = parentFn
+  this.block = block
+  this.slots = slots
+}
 
 function BasicLogger () {
   this.info = console.log
@@ -82,7 +93,7 @@ AST.Root.prototype.emitToFile = function (outFile, opts) {
   // Compile ourselves in the main entry function
   this.buildMain(ctx, mainFunc, mainEntry)
 
-  // ctx.module.dump()
+  ctx.module.dump()
 
   // Verify the module to make sure that nothing's amiss before we hand it
   // of to the bitcode compiler
@@ -159,20 +170,21 @@ function compileBlock (ctx, block, parentFn, preStatementsCb) {
   }
 }
 
-function assertInstanceOf (value, type, message) {
-  if (!(value instanceof type)) {
-    if (!message) {
-      message = 'Incorrect type; expected '+type.name+', got '+value.constructor.name
-    }
-    throw new Error(message)
-  }
-}
+// Look up the type and Slots for a given name; begins search from the passed
+// block-context. Returns a 2-tuple of Slots and Type.
+function getTypeAndSlotsForName (ctx, blockCtx, name, foundCb) {
+  // Keep tracking of the scope of the beginning of the chain
+  var outermostScope = null
+  var type = blockCtx.block.scope.get(name, function (scope, _type) {
+    outermostScope = scope
+  })
 
-function BlockContext (ctx, parentFn, block, slots) {
-  this.ctx   = ctx
-  this.fn    = parentFn
-  this.block = block
-  this.slots = slots
+  // Finally look up the Slots for the outermost scope that the name belongs to
+  var slots = ctx.slotsMap[outermostScope.id]
+  if (!slots) {
+    throw new Error("Couldn't find slots for scope #"+outermostScope.id)
+  }
+  return [slots, type]
 }
 
 AST.Assignment.prototype.compile = function (ctx, blockCtx) {
@@ -262,6 +274,7 @@ AST.Literal.prototype.compileToValue = function (ctx, blockCtx) {
   }
 }
 
+/*
 // FIXME: This can probably be completely removed now that ClosingScope is
 //        integrated into the type-system and preventing closure overreach
 //        at that earlier stage of compilation.
@@ -307,38 +320,7 @@ function computeScopePath (blockCtx, outermostScope) {
   }
   return path
 }
-
-// Look up the type and Slots for a given name; beings search from the passed
-// block-context. Returns a 2-tuple of Slots and Type.
-function getTypeAndSlotsForName (ctx, blockCtx, name, foundCb) {
-  // Keep tracking of the scope of the beginning of the chain
-  var outermostScope = null
-  var type = blockCtx.block.scope.get(name, function (scope, _type) {
-    outermostScope = scope
-  })
-
-  /*
-  // Outermost scope will be the first in `scopePath`; innermost scope
-  // will be the last in `scopePath`.
-  var scopePath = computeScopePath(blockCtx, outermostScope)
-
-  // Sanity-check the `scopePath` to make sure it's outermost scope is the
-  // same as the one we got from the Scope class' scope resolver.
-  if (scopePath[0] !== outermostScope) {
-    throw new Error('Invalid scope path: first scope does not match outermost scope')
-  }
-
-  // Check this scope path to make sure we didn't bust any closures
-  checkScopePathForClosureOverreach(scopePath)
-  */
-
-  // Finally look up the Slots for the outermost scope that the name belongs to
-  var slots = ctx.slotsMap[outermostScope.id]
-  if (!slots) {
-    throw new Error("Couldn't find slots for scope #"+outermostScope.id)
-  }
-  return [slots, type]
-}
+*/
 
 AST.Chain.prototype.compileInstanceMethodCall = function (ctx, blockCtx, receiver, receiverType, instanceMethod, callNode) {
   assertInstanceOf(receiver, Buffer)
@@ -361,6 +343,24 @@ AST.Chain.prototype.compileInstanceMethodCall = function (ctx, blockCtx, receive
   var returnValue = ctx.builder.buildCall(nativeFunction.fn.ptr, args, ''),
       returnType  = new types.Instance(instanceMethod.ret)
   return [returnType, returnValue]
+}
+
+function buildPointerCastIfNecessary (ctx, value, desiredType) {
+  var valueType = TypeOf(value)
+  // Compare the type strings
+  var vts = PrintTypeToString(valueType),
+      dts = PrintTypeToString(desiredType),
+      typesAreEqual = (vts === dts)
+  // If the types aren't the same then we'll recast
+  if (!typesAreEqual) {
+    if (ctx.noCastValuePointers) {
+      throw new ICE('Value type different than the one desired and no-cast-value-pointers is true: '+vts+' -> '+dts)
+    }
+    // Re-cast the value
+    return ctx.builder.buildPointerCast(value, desiredType, '')
+  }
+  // Didn't need to re-cast the value
+  return value
 }
 
 AST.Chain.prototype.compileToValue = function (ctx, blockCtx) {
@@ -435,30 +435,6 @@ AST.Chain.prototype.compileToValue = function (ctx, blockCtx) {
     }
   }
   return itemValue
-}
-
-function buildPointerCastIfNecessary (ctx, value, desiredType) {
-  var valueType = TypeOf(value)
-  // Compare the type strings
-  var vts = PrintTypeToString(valueType),
-      dts = PrintTypeToString(desiredType),
-      typesAreEqual = (vts === dts)
-  // If the types aren't the same then we'll recast
-  if (!typesAreEqual) {
-    if (ctx.noCastValuePointers) {
-      throw new ICE('Value type different than the one desired and no-cast-value-pointers is true: '+vts+' -> '+dts)
-    }
-    // Re-cast the value
-    return ctx.builder.buildPointerCast(value, desiredType, '')
-  }
-  // Didn't need to re-cast the value
-  return value
-}
-
-function nativeTypesEqual (a, b) {
-  var aTypeString = PrintTypeToString(a),
-      bTypeString = PrintTypeToString(b)
-  return (a === b)
 }
 
 AST.Chain.prototype.compile = function (ctx, blockCtx) {
@@ -638,6 +614,7 @@ AST.New.prototype.compile = function (ctx, blockCtx) {
   this.compileToValue(ctx, blockCtx)
 }
 
+// TODO: Maybe move this into `AST.New.prototype`?
 function findInitializer (nativeObject, argsNodes) {
   var type         = nativeObject.type,
       initializers = nativeObject.initializers
@@ -664,12 +641,6 @@ function findInitializer (nativeObject, argsNodes) {
     }
   }
   throw new ICE('Initializer not found')
-}
-
-// Return the size of a type in bytes for use in `malloc` and friends
-function getAllocSizeOfType (ctx, ty) {
-  var targetData = ctx.targetData
-  return LLVM.Library.LLVMABISizeOfType(targetData, ty)
 }
 
 AST.New.prototype.compileToValue = function (ctx, blockCtx) {
@@ -701,25 +672,6 @@ AST.New.prototype.compileToValue = function (ctx, blockCtx) {
   ctx.builder.buildCall(initFn, argValues, '')
   // Return the pointer to the actual object
   return objPtr
-}
-
-function compileTruthyTest (ctx, blockCtx, expr) {
-  var value    = expr.compileToValue(ctx, blockCtx),
-      instance = expr.type
-  // Can only truthy-test instances
-  assertInstanceOf(instance, types.Instance)
-  type = instance.type
-  switch (type.constructor) {
-  case types.String:
-    var nullStringPtr = LLVM.Library.LLVMConstNull(Int8PtrType)
-    // Compare the string pointer to the NULL pointer
-    return ctx.builder.buildICmp(LLVM.Library.LLVMIntNE, value, nullStringPtr, '')
-  case types.Number:
-    var zeroInteger = LLVM.Library.LLVMConstInt(Int64Type, 0, true)
-    return ctx.builder.buildICmp(LLVM.Library.LLVMIntNE, value, zeroInteger, '')
-  default:
-    throw new ICE('Cannot compile to truthy-testable value: '+type.constructor.name)
-  }
 }
 
 var ifCounter = 1
@@ -812,16 +764,5 @@ AST.If.prototype.compileConditionBlock = function (ctx, parentFn, blockNode, blo
   if (!lastInstrTerm && contBlockPtr !== null) {
     ctx.builder.buildBr(contBlockPtr)
   }
-}
-
-function isLastInstructionTerminator (bb) {
-  var lastInstr = LLVM.Library.LLVMGetLastInstruction(bb)
-  // Do nothing if this block is empty
-  if (lastInstr.isNull()) {
-    return false
-  }
-  // Get the opcode and check if it's a terminator
-  var lastInstrOpcode = LLVM.Library.LLVMGetInstructionOpcode(lastInstr)
-  return (LLVM.Library.TerminatorInstructions.indexOf(lastInstrOpcode) !== -1)
 }
